+++ internals/webpack/webpack.base.babel.js
@@ -166,7 +166,8 @@ module.exports = options => ({
       '@themes': path.resolve(__dirname, '../../app/themes'),
       '@images': path.resolve(__dirname, '../../app/images'),
       '@hooks': path.resolve(__dirname, '../../app/hooks'),
-      moment$: path.resolve(__dirname, '../../node_modules/moment/moment.js')
+      moment$: path.resolve(__dirname, '../../node_modules/moment/moment.js'),
+      '@ant-design/icons/lib/dist$': path.resolve(__dirname, './app/icons.js')
     },
     extensions: ['.js', '.jsx', '.react.js'],
     mainFields: ['browser', 'jsnext:main', 'main']
